requirements.txt
numpy
matplotlib
scipy
tensorflow-gpu==2.2.0
opencv-python
sklearn